Refactor suggested labels to separate view
