+++ WalletWasabi.Fluent/ViewModels/Wallets/Receive/ReceiveViewModel.cs
@@ -1,18 +1,13 @@
-using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Linq;
-using System.Reactive;
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
 using System.Windows.Input;
-using DynamicData;
 using DynamicData.Binding;
 using ReactiveUI;
 using WalletWasabi.Blockchain.Analysis.Clustering;
 using WalletWasabi.Blockchain.Keys;
-using WalletWasabi.Fluent.Helpers;
 using WalletWasabi.Fluent.ViewModels.Navigation;
+using WalletWasabi.Fluent.ViewModels.Wallets.Labels;
 using WalletWasabi.Wallets;
 
 namespace WalletWasabi.Fluent.ViewModels.Wallets.Receive
@@ -27,71 +22,28 @@ namespace WalletWasabi.Fluent.ViewModels.Wallets.Receive
 	public partial class ReceiveViewModel : RoutableViewModel
 	{
 		private readonly Wallet _wallet;
-		private readonly SourceList<SuggestionLabelViewModel> _suggestionLabels;
-		private readonly ObservableCollectionExtended<SuggestionLabelViewModel> _suggestionLabelResults;
-		private readonly ObservableCollectionExtended<string> _labels;
-		[AutoNotify] private HashSet<string> _suggestions;
 		[AutoNotify] private bool _isExistingAddressesButtonVisible;
+		[AutoNotify(SetterModifier = AccessModifier.Private)] private SuggestionLabelsViewModel _suggestionLabels;
 
 		public ReceiveViewModel(Wallet wallet)
 		{
 			_wallet = wallet;
-			_labels = new ObservableCollectionExtended<string>();
-			var allLabels = WalletHelpers.GetLabels();
-
-			var mostUsedLabels = allLabels.GroupBy(x => x)
-				.Select(x => new
-				{
-					Label = x.Key,
-					Count = x.Count()
-				})
-				.OrderByDescending(x => x.Count)
-				.ToList();
-
-			_suggestions = mostUsedLabels.Select(x => x.Label).ToHashSet();
-
-			_suggestionLabels = new SourceList<SuggestionLabelViewModel>();
-			_suggestionLabelResults = new ObservableCollectionExtended<SuggestionLabelViewModel>();
-
-			_suggestionLabels.AddRange(
-				mostUsedLabels.Select(x => new SuggestionLabelViewModel(x.Label, x.Count)));
-
-			var suggestionLabelsFilter = this.WhenAnyValue(x => x.Labels).Select(_ => Unit.Default)
-				.Merge(Observable.FromEventPattern(Labels, nameof(Labels.CollectionChanged)).Select(_ => Unit.Default))
-				.Select(SuggestionLabelsFilter);
-
-			_suggestionLabels
-				.Connect()
-				.Filter(suggestionLabelsFilter)
-				.Sort(SortExpressionComparer<SuggestionLabelViewModel>.Descending(x => x.Count))
-				.Top(3)
-				.ObserveOn(RxApp.MainThreadScheduler)
-				.Bind(_suggestionLabelResults)
-				.Subscribe();
-
 			SetupCancel(enableCancel: true, enableCancelOnEscape: true, enableCancelOnPressed: true);
 
 			EnableBack = false;
 
-			NextCommand = ReactiveCommand.Create(OnNext, _labels.ToObservableChangeSet().Select(_ => _labels.Count > 0));
+			_suggestionLabels = new SuggestionLabelsViewModel();
+
+			NextCommand = ReactiveCommand.Create(OnNext, _suggestionLabels.Labels.ToObservableChangeSet().Select(_ => _suggestionLabels.Labels.Count > 0));
 
 			ShowExistingAddressesCommand = ReactiveCommand.Create(OnShowExistingAddresses);
 		}
 
 		public ICommand ShowExistingAddressesCommand { get; }
 
-		public ObservableCollection<SuggestionLabelViewModel> SuggestionLabelResults => _suggestionLabelResults;
-
-		public ObservableCollection<string> Labels => _labels;
-
-		private Func<SuggestionLabelViewModel, bool> SuggestionLabelsFilter(Unit unit)
-		{
-			return suggestionLabel => !_labels.Contains(suggestionLabel.Label);
-		}
-
 		private void OnNext()
 		{
-			var newKey = _wallet.KeyManager.GetNextReceiveKey(new SmartLabel(Labels), out bool minGapLimitIncreased);
+			var newKey = _wallet.KeyManager.GetNextReceiveKey(new SmartLabel(SuggestionLabels.Labels), out bool minGapLimitIncreased);
 
 			if (minGapLimitIncreased)
 			{
@@ -102,14 +54,14 @@ namespace WalletWasabi.Fluent.ViewModels.Wallets.Receive
 				// TODO: notification
 			}
 
-			Labels.Clear();
+			_suggestionLabels.Labels.Clear();
 
 			Navigate().To(new ReceiveAddressViewModel(_wallet, newKey));
 		}
 
 		private void OnShowExistingAddresses()
 		{
-			Navigate().To(new ReceiveAddressesViewModel(_wallet, Suggestions));
+			Navigate().To(new ReceiveAddressesViewModel(_wallet, _suggestionLabels.Suggestions));
 		}
 
 		protected override void OnNavigatedTo(bool isInHistory, CompositeDisposable disposable)
