+++ WalletWasabi.Fluent/Views/Wallets/Labels/SuggestionLabelsView.axaml.cs
@@ -1,0 +1,18 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WalletWasabi.Fluent.Views.Wallets.Labels
+{
+	public class SuggestionLabelsView : UserControl
+	{
+		public SuggestionLabelsView()
+		{
+			InitializeComponent();
+		}
+
+		private void InitializeComponent()
+		{
+			AvaloniaXamlLoader.Load(this);
+		}
+	}
+}
