+++ WalletWasabi.Fluent/Views/Wallets/Receive/ReceiveView.axaml
@@ -7,6 +7,7 @@
              xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors"
              xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
              xmlns:receive="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets.Receive"
+             xmlns:labels="clr-namespace:WalletWasabi.Fluent.Views.Wallets.Labels"
              x:DataType="receive:ReceiveViewModel"
              x:CompileBindings="True"
              x:Class="WalletWasabi.Fluent.Views.Wallets.Receive.ReceiveView">
@@ -25,7 +26,10 @@
           <TextBlock TextWrapping="Wrap" Margin="0 0 10 0"  Text="Who will know this receive address belongs to you?" HorizontalAlignment="Left"
                  FontWeight="Normal" />
         </DockPanel>
-        <c:TagsBox x:Name="LabelsTbx" Items="{Binding Labels}"
+        <c:TagsBox x:Name="LabelsTbx"
+                   DataContext="{Binding SuggestionLabels}"
+                   SetAddTagAction="{Binding SetAddTag}"
+                   Items="{Binding Labels}"
                    TagSeparator=","
                    SuggestionsAreCaseSensitive="True"
                    RestrictInputToSuggestions="False"
@@ -40,25 +44,7 @@
             <behaviors:FocusOnAttachedBehavior />
           </i:Interaction.Behaviors>
         </c:TagsBox>
-        <ItemsControl Items="{Binding SuggestionLabelResults}"
-                      ClipToBounds="False">
-          <ItemsControl.ItemsPanel>
-            <ItemsPanelTemplate>
-              <StackPanel Orientation="Horizontal" Spacing="2" />
-            </ItemsPanelTemplate>
-          </ItemsControl.ItemsPanel>
-          <ItemsControl.ItemTemplate>
-            <DataTemplate x:DataType="receive:SuggestionLabelViewModel">
-              <Button Classes="option"
-                      Margin="0,0,4,0"
-                      Focusable="False"
-                      Content="{Binding Label}"
-                      Command="{Binding #LabelsTbx.AddTag}"
-                      CommandParameter="{Binding Label}"
-                      x:CompileBindings="False"/>
-            </DataTemplate>
-          </ItemsControl.ItemTemplate>
-        </ItemsControl>
+        <labels:SuggestionLabelsView DataContext="{Binding SuggestionLabels}"/>
       </StackPanel>
     </DockPanel>
   </c:ContentArea>
